chore(react): rename directory name
