+++ packages/react/src/examples/playground/05-React/Slots/SlotsSample.tsx
