+++ packages/react/src/examples/playground/index.ts
@@ -1,5 +1,5 @@
-export { Basics } from './06-React/Basics/Basics';
-export { ReactEvents } from './06-React/ReactEvents/ReactEvents';
-export { CustomEvents } from './06-React/CustomEvents/CustomEvents';
-export { Slots } from './06-React/Slots/Slots';
-export { Refs } from './06-React/Refs/Refs';
+export { Basics } from './05-React/Basics/Basics';
+export { ReactEvents } from './05-React/ReactEvents/ReactEvents';
+export { CustomEvents } from './05-React/CustomEvents/CustomEvents';
+export { Slots } from './05-React/Slots/Slots';
+export { Refs } from './05-React/Refs/Refs';
